requirements.txt
PyYAML==6.0.1
beautifulsoup4==4.12.3
httpx[brotli]==0.27.0
loguru==0.7.2
pydantic==2.7.3
tenacity==8.3.0
